requirements.txt
streamlit==1.44.1
opencv-python==4.9.0.80
mediapipe==0.10.21
numpy==1.26.4
pandas==2.2.3
Pillow==11.2.1
joblib==1.3.2
scikit-learn==1.6.1 